+++ src/components/Chat/ChatMessage.tsx
@@ -2,19 +2,34 @@ import * as React from 'react'
 import * as PropTypes from 'prop-types'
 import * as cx from 'classnames'
 
-import { childrenExist, createShorthandFactory, customPropTypes, UIComponent } from '../../lib'
-import { ComponentVariablesInput, IComponentPartStylesInput } from '../../../types/theme'
+import {
+  childrenExist,
+  createShorthandFactory,
+  customPropTypes,
+  UIComponent,
+  IRenderResultConfig,
+  pxToRem,
+} from '../../lib'
+import {
+  ComponentVariablesInput,
+  IComponentPartStylesInput,
+  ComponentPartStyle,
+} from '../../../types/theme'
 import { Extendable, ReactChildren, ItemShorthand } from '../../../types/utils'
 import Avatar from '../Avatar'
+import Layout from '../Layout'
+import Text from '../Text'
 
 export interface IChatMessageProps {
   as?: any
+  author?: string
   avatar?: ItemShorthand
   children?: ReactChildren
   className?: string
   content?: any
   mine?: boolean
   styles?: IComponentPartStylesInput
+  timestamp?: string
   variables?: ComponentVariablesInput
 }
 
@@ -27,6 +42,9 @@ class ChatMessage extends UIComponent<Extendable<IChatMessageProps>, any> {
 
   static propTypes = {
     as: customPropTypes.as,
+
+    /** Author of the message. */
+    author: PropTypes.string,
 
     /** Chat messages can have an avatar */
     avatar: customPropTypes.itemShorthand,
@@ -46,18 +64,23 @@ class ChatMessage extends UIComponent<Extendable<IChatMessageProps>, any> {
     /** Custom styles to be applied for component. */
     styles: PropTypes.oneOfType([PropTypes.object, PropTypes.func]),
 
+    /** Timestamp of the message. */
+    timestamp: PropTypes.string,
+
     /** Custom variables to be applied for component. */
     variables: PropTypes.oneOfType([PropTypes.object, PropTypes.func]),
   }
 
   static handledProps = [
     'as',
+    'author',
     'avatar',
     'children',
     'className',
     'content',
     'mine',
     'styles',
+    'timestamp',
     'variables',
   ]
 
@@ -65,31 +88,67 @@ class ChatMessage extends UIComponent<Extendable<IChatMessageProps>, any> {
     as: 'li',
   }
 
-  renderComponent({ ElementType, classes, rest, styles, variables }) {
-    const { avatar, children, content, mine } = this.props
+  renderComponent({
+    ElementType,
+    classes,
+    rest,
+    styles,
+    variables,
+  }: IRenderResultConfig<IChatMessageProps>) {
+    const { as, avatar, children, mine } = this.props
 
     return childrenExist(children) ? (
       <ElementType {...rest} className={cx(classes.root, classes.content)}>
         {children}
       </ElementType>
     ) : (
-      <ElementType {...rest} className={classes.root}>
-        {!mine && this.renderAvatar(avatar, styles, variables)}
-        <div className={classes.content}>{content}</div>
-        {mine && this.renderAvatar(avatar, styles, variables)}
-      </ElementType>
+      <Layout
+        as={as}
+        {...rest}
+        className={classes.root}
+        start={!mine && this.renderAvatar(avatar, styles.avatar, variables)}
+        main={this.renderContent(classes.content)}
+        end={mine && this.renderAvatar(avatar, styles.avatar, variables)}
+      />
     )
   }
 
+  private renderContent = (contentClass: string) => {
+    const { author, content, mine, timestamp } = this.props
+
+    return (
+      <Layout
+        className={contentClass}
+        vertical
+        start={
+          <>
+            {!mine && author && this.renderText(author)}
+            {timestamp && this.renderText(timestamp, true)}
+          </>
+        }
+        main={content}
+      />
+    )
+  }
+
+  private renderText = (content: string, timestamp?: boolean) => (
+    <Text
+      timestamp={timestamp}
+      content={content}
+      size="sm"
+      styles={{ root: { marginRight: pxToRem(10) } }}
+    />
+  )
+
   private renderAvatar = (
     avatar: ItemShorthand,
-    styles: IComponentPartStylesInput,
+    avatarStyles: ComponentPartStyle,
     variables: ComponentVariablesInput,
   ) =>
     avatar &&
     Avatar.create(avatar, {
       defaultProps: {
-        styles: { root: styles.avatar },
+        styles: { root: avatarStyles },
         variables: variables.avatar,
       },
     })